+++ noodles-sam/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+  * sam/record/data/field/tag: Compare inner tag.
+
 ## 0.40.0 - 2023-08-24
 
 ### Added